% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDiagnosticsSettings.R
\name{createDataDiagnosticsSettings}
\alias{createDataDiagnosticsSettings}
\title{createDataDiagnosticsSettings}
\usage{
createDataDiagnosticsSettings(
  analysisId,
  analysisName = "analysis",
  minAge = NULL,
  maxAge = NULL,
  genderConceptIds = c(8507, 8532),
  raceConceptIds = NULL,
  ethnicityConceptIds = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredDurationDays = 365,
  requiredDomains = c("condition", "drug"),
  desiredDomains = NULL,
  requiredVisits = NULL,
  desiredVisits = NULL,
  targetName = "target",
  targetConceptIds = NULL,
  targetUseDrugEra = F,
  comparatorName = NULL,
  comparatorConceptIds = NULL,
  comparatorUseDrugEra = F,
  indicationName = NULL,
  indicationConceptIds = NULL,
  indicationUseDrugEra = F,
  includeIndicationInCalc = F,
  outcomeName = NULL,
  outcomeConceptIds = NULL,
  outcomeUseDrugEra = F
)
}
\arguments{
\item{analysisId}{The identifier for the specific analysis.}

\item{analysisName}{The name for the specific analysis.}

\item{minAge}{The minimum age for patients included in the analysis.
Default is the lowest age available in the database.}

\item{maxAge}{The minimum age for patients included in the analysis.
Default is the highest age available in the database.}

\item{genderConceptIds}{The required sex(es) at birth for patients included in the analysis, expressed as a vector of concepts.
List the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Gender&standardConcept=Standard&page=1&pageSize=15&query=).
Default is c(8507, 8532) for male and female.}

\item{raceConceptIds}{The required races for patients included in the analysis, expressed as a vector of concepts, e.g. c(8515,8527).
If no restriction leave blank, else list the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Race&standardConcept=Standard&page=1&pageSize=15&query=).}

\item{ethnicityConceptIds}{The required ethnicities for patients included in the analysis, expressed as a vector of concepts, e.g. c(38003564,38003563)
If no restriction leave blank, else list the standard OMOP concepts for the required values as found here (https://athena.ohdsi.org/search-terms/terms?domain=Ethnicity&standardConcept=Standard&page=1&pageSize=15&query=).}

\item{studyStartDate}{The start date of the analysis. Date format it 'YYYYMM'. If no restriction leave blank.}

\item{studyEndDate}{The end date of the analysis. Date format it 'YYYYMM'. If no restriction leave blank.}

\item{requiredDurationDays}{The minimum required follow-up time in days for patients included in the analysis.}

\item{requiredDomains}{The data domains required for ALL patients in the analysis, expressed as a character vector.
Valid values are condition,drug,device,measurement,measurementValues,death,procedure,observation. Default is c("condition", "drug")}

\item{desiredDomains}{The domains required for SOME patients in the analysis. Meaning you need the dataset to have these data but each person is not
required to have a record. This is most often used to identify the presence of outcomes of interest. Valid values are condition,drug,device,measurement,measurementValues,death,procedure,observation.
If no restriction leave blank.}

\item{requiredVisits}{The visits required for ALL patients in the study, expressed as a character vector, e.g. c("IP","OP"). Valid values are IP,OP,ER. If no restriction leave blank.}

\item{desiredVisits}{The visits required for SOME patients in the study, expressed as a character vector.
Meaning you need the dataset to have these data but each person is not required to have a record. This is most often used to identify the presence of outcomes of interest. Valid values are IP,OP,ER.}

\item{targetName}{The name of the target of interest.}

\item{targetConceptIds}{A vector containing the required target concepts.}

\item{targetUseDrugEra}{Should the DRUG_ERA table be used for identifying targets? Set to T only if targetConceptIds param is a list of RxNorm ingredients.}

\item{comparatorName}{The name of the comparator of interest.}

\item{comparatorConceptIds}{A vector containing the required comparator concepts.}

\item{comparatorUseDrugEra}{Should the DRUG_ERA table be used for identifying comparators? Set to T only if comparatorConceptIds param is a list of RxNorm ingredients.}

\item{indicationName}{The name of the indication of interest.}

\item{indicationConceptIds}{A vector containing the required indication concepts.}

\item{indicationUseDrugEra}{Should the DRUG_ERA table be used for identifying the indication? Set to T only if indicationConceptIds param is a list of RxNorm ingredients.}

\item{includeIndicationInCalc}{A T/F object indicating whether the proportion of people with the indication concepts should be included in the estimated sample size calculation. Default is FALSE.}

\item{outcomeName}{The name of the outcome of interest.}

\item{outcomeConceptIds}{A vector containing the required outcome concepts.}

\item{outcomeUseDrugEra}{Should the DRUG_ERA table be used for identifying the outcome? Set to T only if outcomeConceptIds param is a list of RxNorm ingredients.}
}
\description{
createDataDiagnosticsSettings
}
