% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDiagnosticsSummary.R
\name{createDataDiagnosticsSummary}
\alias{createDataDiagnosticsSummary}
\title{createDataDiagnosticsSummary}
\usage{
createDataDiagnosticsSummary(dbDiagnosticsResults)
}
\arguments{
\item{dbDiagnosticsResults}{A dataframe with the full data diagnostics output from the \code{executeDbDiagnostics} function.}
}
\description{
createDataDiagnosticsSummary
}
