% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePDF.R
\name{createPdfReport}
\alias{createPdfReport}
\title{createPdfReport}
\usage{
createPdfReport(
  results,
  file,
  reportName = "User Defined Report Name",
  userNotes = NULL,
  analysisSettings = NULL
)
}
\arguments{
\item{results}{DbDiagnostics results table}

\item{file}{Name and location for generated report, example: "./report.pdf"}
}
\description{
createPdfReport
}
