% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbDiagnostics.R
\docType{package}
\name{DbDiagnostics-package}
\alias{DbDiagnostics}
\alias{DbDiagnostics-package}
\title{DbDiagnostics: OMOP CDM Database Diagnostics Utility}
\description{
Functions for extracting profile data for an OMOP CDM database and determining if the database is appropriate for individual use cases.
}
\author{
\strong{Maintainer}: Clair Blacketer \email{mblacke@its.jnj.com}

Authors:
\itemize{
  \item Frank DeFalco
}

}
\keyword{internal}
