% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeDbDiagnostics.R
\name{executeDbDiagnostics}
\alias{executeDbDiagnostics}
\title{executeDbDiagnostics}
\usage{
executeDbDiagnostics(
  connectionDetails,
  resultsDatabaseSchema,
  resultsTableName,
  outputFolder = getwd(),
  dataDiagnosticsSettings
)
}
\arguments{
\item{connectionDetails}{A connectionDetails object for connecting to the database containing the DbProfile results}

\item{resultsDatabaseSchema}{The fully qualified database name of the results schema where the DbProfile results are housed}

\item{resultsTableName}{The name of the table in the results schema with the DbProfile results}

\item{outputFolder}{Results will be written to this directory, default = getwd()}

\item{dataDiagnosticsSettings}{A list of settings created from DataDiagnostics::createDataDiagnosticsSettings() function}
}
\description{
executeDbDiagnostics
}
