% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeDbProfile.R
\name{executeDbProfile}
\alias{executeDbProfile}
\title{executeDbProfile}
\usage{
executeDbProfile(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema,
  cdmSourceName,
  outputFolder = getwd(),
  cdmVersion = "5.3",
  overwriteAchilles = FALSE,
  minCellCount = 5,
  addDQD = TRUE,
  tableCheckThresholds = "default",
  fieldCheckThresholds = "default",
  conceptCheckThresholds = "default"
)
}
\arguments{
\item{connectionDetails}{A connectionDetails object for connecting to the CDM database}

\item{cdmDatabaseSchema}{The fully qualified database name of the CDM schema}

\item{resultsDatabaseSchema}{The fully qualified database name of the results schema}

\item{vocabDatabaseSchema}{The fully qualified database name of the vocabulary schema}

\item{cdmSourceName}{The name of the CDM data source}

\item{outputFolder}{Results will be written to this directory, Default = getwd()}

\item{cdmVersion}{The CDM version to target for the data source. Default = "5.3"}

\item{overwriteAchilles}{Specify if existing achilles results tables should be overwritten, Default=FALSE}

\item{minCellCount}{Minimum cell count to allow in analyses. Default = 0}

\item{addDQD}{Specify if DQD should be run. Default = TRUE}

\item{tableCheckThresholds}{OPTIONAL Location of the custom threshold file for the DQD table checks. In absence of this parameter the default thresholds are used.}

\item{fieldCheckThresholds}{OPTIONAL Location of the custom threshold file for the DQD field checks. In absence of this parameter the default thresholds are used.}

\item{conceptCheckThresholds}{OPTIONAL Location of the custom threshold file for the DQD concept checks. In absence of this parameter the default thresholds are used.}
}
\description{
executeDbProfile
}
