% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeDbDiagnostics.R
\name{executeDbDiagnostics}
\alias{executeDbDiagnostics}
\title{executeDbDiagnostics}
\usage{
executeDbDiagnostics(
  connectionDetails,
  resultsDatabaseSchema = "dp_temp",
  resultsTableName = "dp_achilles_results_augmented",
  outputFolder = getwd(),
  dataDiagnosticsSettingsList
)
}
\arguments{
\item{connectionDetails}{A connectionDetails object for connecting to the database containing the DbProfile results}

\item{resultsDatabaseSchema}{The fully qualified database name of the results schema where the DbProfile results are housed. Default is "dp_temp".}

\item{resultsTableName}{The name of the table in the results schema with the DbProfile results. Default is "dp_achilles_results_augmented."}

\item{outputFolder}{Results will be written to this directory, default = getwd()}

\item{dataDiagnosticsSettingsList}{A list of settings objects, each created from DataDiagnostics::createDataDiagnosticsSettings() function and each representing one analysis.}
}
\description{
executeDbDiagnostics
}
